Break out constants file to fix circular dependency
