+++ packages/atxp-client/src/baseConstants.ts
@@ -1,0 +1,1 @@
+export const USDC_CONTRACT_ADDRESS_BASE = "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913"; // USDC on Base mainnet
