+++ packages/atxp-client/src/basePaymentMaker.insufficientFunds.test.ts
@@ -2,7 +2,7 @@ import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { BasePaymentMaker } from './basePaymentMaker.js';
 import { InsufficientFundsError, PaymentNetworkError } from './types.js';
 import { BigNumber } from 'bignumber.js';
-import { USDC_CONTRACT_ADDRESS_BASE } from './baseAccount.js';
+import { USDC_CONTRACT_ADDRESS_BASE } from './baseConstants.js';
 
 // Mock viem functions
 vi.mock('viem', () => ({
